+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Moved `useTree(tree)` module hook function to `tree.moduleHooks()` method
 * Removed undocumented and unuseful `exportAsString` module hook
+* Changed `DevServer.files` to match `FileTree.files`
+* Changed `generateFiles` file map to match `FileTree.files`
 
 ## 2.6.1